remove unused code, highlight current beat
