+++ pages/index.tsx
@@ -1,18 +1,25 @@
 import Head from "next/head";
 import { Suspense } from "react";
 import dynamic from "next/dynamic";
+import notesAtom, { TNotes } from "../src/lib/store/notesAtom";
+import NoteRow from "../src/components/notes/NoteRow";
+import { useRecoilValue } from "recoil";
 
-// const SynthWrapper = dynamic(
-//   () => import("../src/components/synth/SynthWrapper"),
-//   {
-//     ssr: false,
-//   }
-// );
 const Machine = dynamic(() => import("../src/components/machine/MachineView"), {
   ssr: false,
 });
 
 export default function Home() {
+  const notes = useRecoilValue(notesAtom);
+  const noteTable = notes.map((note, index) => (
+    <NoteRow
+      index={index}
+      key={Math.random()}
+      activeNotes={note.activeNotes}
+      waveShape={note.waveShape}
+      note={note.note}
+    />
+  ));
   return (
     <>
       <Head>
@@ -21,9 +28,7 @@ export default function Home() {
       <div className="flex justify-around">
         <h1>Synth App</h1>
       </div>
-      {/* <Suspense> */}
-      {/*   <SynthWrapper /> */}
-      {/* </Suspense> */}
+      <div className="flex-col w-full">{noteTable}</div>
       <Suspense>
         <Machine></Machine>
       </Suspense>
