+++ src/components/notes/Note/index.tsx
@@ -1,6 +1,6 @@
-import { useRecoilState } from "recoil";
+import { useRecoilState, useRecoilValue } from "recoil";
 import activeNotesState from "../../../lib/store/activeNotesState";
-
+import noteIsActive from "../../../lib/store/noteIsActive";
 export interface NoteProps {
   note: string;
   index: number;
@@ -10,6 +10,7 @@ const Note = ({ note, indexRow, index }: NoteProps) => {
   const [activeNotes, setActiveNotes] = useRecoilState(
     activeNotesState(indexRow)
   );
+  const noteIsActiveState = useRecoilValue(noteIsActive(index));
   const handleUpdateIsActive = () => {
     const state = [...activeNotes];
     state[index] = !activeNotes[index];
@@ -17,12 +18,13 @@ const Note = ({ note, indexRow, index }: NoteProps) => {
   };
 
   const styleActive = "bg-gray-500";
+  const styleCurrentBeat = "opacity-75";
 
   return (
     <button
       className={`border border-black w-full ${
         activeNotes[index] ? styleActive : null
-      }`}
+      } ${noteIsActiveState ? styleCurrentBeat : null}`}
       onClick={handleUpdateIsActive}
     >{`${note} ${index}`}</button>
   );
