--- src/components/synth/SynthWrapper.tsx
@@ -1,67 +1,0 @@
-import * as Tone from "tone";
-import NoteNode from "./NoteNode";
-import { useRecoilValue, useRecoilState, useSetRecoilState } from "recoil";
-import noteTableState from "./store/notesAtom";
-import tableAtom from "./store/tableAtom";
-import currentRowAtom from "./store/currentRowAtom";
-
-const SynthWrapper = ({}) => {
-  const noteTable = useRecoilValue(noteTableState);
-  const table = useRecoilValue(tableAtom);
-  const setCurrentRow = useSetRecoilState(currentRowAtom);
-
-  const handlePlaySynth = () => {
-    const synth = new Tone.Synth().toDestination();
-    new Tone.Sequence(
-      (time, note) => {
-        synth.triggerAttackRelease(note, "8n", time);
-      },
-      noteTable.map((note) => `${note.note}${note.octave}`)
-    ).start(0);
-    Tone.Transport.timeSignature = [9, 8];
-    Tone.Transport.bpm.value = 120;
-    // Tone.Transport.scheduleRepeat(() => {
-    //   setCurrentRow((row) => (row < table.width ? row + 1 : 1));
-    // }, "8n");
-    //
-    // create a second sequence with a new instrument
-    const synth2 = new Tone.Synth().toDestination();
-    new Tone.Sequence(
-      (time, note) => {
-        synth2.triggerAttackRelease(note, "8n", time);
-      },
-      ["C4"]
-    ).start(0);
-    Tone.Transport.timeSignature = [9, 8];
-    Tone.Transport.bpm.value = 120;
-    Tone.start();
-    Tone.Transport.timeSignature = [9, 8];
-    Tone.Transport.start();
-  };
-  const notesRow = [];
-  for (let i = 0; i < table.width; i++) {
-    notesRow.push(<NoteNode key={i} index={i} />);
-  }
-
-  return (
-    <div className="flex justify-around p-2">
-      <button
-        onClick={() => handlePlaySynth()}
-        className="p-2 border border-black text-3xl"
-      >
-        Start
-      </button>
-      <button
-        onClick={() => {
-          Tone.Transport.stop();
-          setCurrentRow(0);
-        }}
-      >
-        Stop
-      </button>
-      {notesRow}
-    </div>
-  );
-};
-
-export default SynthWrapper;
