--- src/components/synth/store/currentRowAtom.ts
@@ -1,8 +1,0 @@
-import { atom } from "recoil";
-
-const currentRowAtom = atom({
-  key: "currentRow",
-  default: 0,
-});
-
-export default currentRowAtom;
